+++ frontend/src/components/chat-message-tool.tsx
@@ -147,7 +147,7 @@ export function ChatMessageTool({
   };
 
   return (
-    <div className="border-r-2 p-8 border-white-300 border-solid">
+    <div className="border-r-2 p-3.5 ml-5 border-white-300 border-solid">
       <HumanValidationDialog
         key={tool.toolCallId}
         threadId={threadId}
@@ -162,7 +162,7 @@ export function ChatMessageTool({
       />
       <div className="flex justify-start">
         <Collapsible open={toolOpen} onOpenChange={setToolOpen}>
-          <div className="flex items-center gap-4">
+          <div className="flex items-center gap-2">
             <CollapsibleTrigger className="hover:scale-105 active:scale-[1.10]">
               <span className="text-sm p-4 truncate border-2 bg-blue-500 rounded-xl">
                 {
@@ -218,7 +218,7 @@ export function ChatMessageTool({
                 </div>
                 {tool?.state === "result" && (
                   <div className="flex flex-col mt-4">
-                    <h1>Result</h1>
+                    <h1>{validated === "rejected" ? "Feedback" : "Result"}</h1>
                     <pre className="text-sm p-2 mt-2 rounded-md overflow-auto bg-gray-100 dark:bg-slate-800">
                       {typeof tool?.result === "string"
                         ? (() => {