+++ CHANGELOG.md
@@ -45,6 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Automatic title generation
 - [frontend] - Collapsible side bar.
 - [frontend] - Add confirmation box on thread deletion.
+- Feedback for HIL tool call refusal.
+- [frontend] - Per message tool toggle.
 - [frontend] - Possibility to dynamically chose available tools.
 - [frontend] - Search bar in tool list.
 - Displaying tool metadata